+++ requests/models.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 """
-requests.system
+requests.models
 ~~~~~~~~~~~~~~~
 
 """
@@ -19,51 +19,62 @@ from .monkeys import Request as _Request, HTTPBasicAuthHandler, HTTPDigestAuthHa
 from .structures import CaseInsensitiveDict
 from .packages.poster.encode import multipart_encode
 from .packages.poster.streaminghttp import register_openers, get_handlers
+from .exceptions import RequestException, AuthenticationError, Timeout, URLRequired, InvalidMethod
 
 
 
 class Request(object):
-    """The :class:`Request` object. It carries out all functionality of
+    """The :class:`Request <models.Request>` object. It carries out all functionality of
     Requests. Recommended interface is with the Requests functions.
     """
 
     _METHODS = ('GET', 'HEAD', 'PUT', 'POST', 'DELETE')
 
     def __init__(self, url=None, headers=dict(), files=None, method=None,
-                 data=dict(), auth=None, cookiejar=None, timeout=None, redirect=True):
+                 data=dict(), auth=None, cookiejar=None, timeout=None,
+                 redirect=True, allow_redirects=False):
 
         socket.setdefaulttimeout(timeout)
 
+        #: Request URL.
         self.url = url
+        #: Dictonary of HTTP Headers to attach to the :class:`Request <models.Request>`.
         self.headers = headers
+        #: Dictionary of files to multipart upload (``{filename: content}``).
         self.files = files
+        #: HTTP Method to use. Available: GET, HEAD, PUT, POST, DELETE.
         self.method = method
+        #: Form or Byte data to attach to the :class:`Request <models.Request>`.
         self.data = dict()
+        #: True if :class:`Request <models.Request>` is part of a redirect chain (disables history
+        #: and HTTPError storage).
         self.redirect = redirect
+        #: Set to True if full redirects are allowed (e.g. re-POST-ing of data at new ``Location``)
+        self.allow_redirects = allow_redirects
 
-        # self.data = {}
         if hasattr(data, 'items'):
             for (k, v) in data.items():
                 self.data.update({
                     k.encode('utf-8') if isinstance(k, unicode) else k:
                     v.encode('utf-8') if isinstance(v, unicode) else v
                 })
-
-        # url encode data if it's a dict
-        if hasattr(data, 'items'):
             self._enc_data = urllib.urlencode(self.data)
         else:
-            self._enc_data = data
+            self._enc_data = self.data = data
 
-
+        #: :class:`Response <models.Response>` instance, containing
+        #: content and metadata of HTTP Response, once :attr:`sent <send>`.
         self.response = Response()
 
         if isinstance(auth, (list, tuple)):
             auth = AuthObject(*auth)
         if not auth:
             auth = auth_manager.get_auth(self.url)
+        #: :class:`AuthObject` to attach to :class:`Request <models.Request>`.
         self.auth = auth
+        #: CookieJar to attach to :class:`Request <models.Request>`.
         self.cookiejar = cookiejar
+        #: True if Request has been sent.
         self.sent = False
 
 
@@ -104,9 +115,6 @@ class Request(object):
 
 
         _handlers.append(HTTPRedirectHandler)
-        # print _handlers
-        # print '^^'
-        # print '!'
 
         if not _handlers:
             return urllib2.urlopen
@@ -128,7 +136,7 @@ class Request(object):
         return opener.open
 
     def _build_response(self, resp):
-        """Build internal Response object from given response."""
+        """Build internal :class:`Response <models.Response>` object from given response."""
 
         def build(resp):
 
@@ -158,14 +166,31 @@ class Request(object):
 
         if self.redirect:
 
-            while 'location' in r.headers:
+            while (
+                ('location' in r.headers) and
+                ((self.method in ('GET', 'HEAD')) or
+                (r.status_code is 303) or
+                (self.allow_redirects))
+            ):
 
                 history.append(r)
 
                 url = r.headers['location']
 
+                # Facilitate for non-RFC2616-compliant 'location' headers
+                # (e.g. '/path/to/resource' instead of 'http://domain.tld/path/to/resource')
+                if not urlparse(url).netloc:
+                    parent_url_components = urlparse(self.url)
+                    url = '%s://%s/%s' % (parent_url_components.scheme, parent_url_components.netloc, url)
+
+                # http://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html#sec10.3.4
+                if r.status_code is 303:
+                    method = 'GET'
+                else:
+                    method = self.method
+
                 request = Request(
-                    url, self.headers, self.files, self.method,
+                    url, self.headers, self.files, method,
                     self.data, self.auth, self.cookiejar, redirect=False
                 )
                 request.send()
@@ -234,6 +259,8 @@ class Request(object):
                 self._build_response(why)
                 if not self.redirect:
                     self.response.error = why
+            except urllib2.URLError, error:
+                raise Timeout if isinstance(error.reason, socket.timeout) else error
             else:
                 self._build_response(resp)
                 self.response.ok = True
@@ -253,19 +280,34 @@ class Request(object):
 
 
 class Response(object):
-    """The :class:`Request` object. All :class:`Request` objects contain a
-    :class:`Request.response <response>` attribute, which is an instance of
-    this class.
+    """The core :class:`Response <models.Response>` object. All
+    :class:`Request <models.Request>` objects contain a
+    :class:`response <models.Response>` attribute, which is an instance
+    of this class.
     """
 
     def __init__(self):
+        #: Raw content of the response, in bytes.
+        #: If ``content-encoding`` of response was set to ``gzip``, the
+        #: response data will be automatically deflated.
         self.content = None
+        #: Integer Code of responded HTTP Status.
         self.status_code = None
+        #: Case-insensitive Dictionary of Response Headers.
+        #: For example, ``headers['content-encoding']`` will return the
+        #: value of a ``'Content-Encoding'`` response header.
         self.headers = CaseInsensitiveDict()
+        #: Final URL location of Response.
         self.url = None
+        #: True if no :attr:`error` occured.
         self.ok = False
+        #: Resulting :class:`HTTPError` of request, if one occured.
         self.error = None
+        #: True, if the response :attr:`content` is cached locally.
         self.cached = False
+        #: A list of :class:`Response <models.Response>` objects from
+        #: the history of the Request. Any redirect responses will end
+        #: up here.
         self.history = []
 
 
@@ -274,22 +316,24 @@ class Response(object):
 
 
     def __nonzero__(self):
-        """Returns true if status_code is 'OK'."""
+        """Returns true if :attr:`status_code` is 'OK'."""
         return not self.error
 
 
     def raise_for_status(self):
-        """Raises stored HTTPError if one exists."""
+        """Raises stored :class:`HTTPError`, if one occured."""
         if self.error:
             raise self.error
 
     def read(self, *args):
+        """Returns :attr:`content`. Used for file-like object compatiblity."""
+
         return self.content
 
 
 
 class AuthManager(object):
-    """Authentication Manager."""
+    """Requests Authentication Manager."""
 
     def __new__(cls):
         singleton = cls.__dict__.get('__singleton__')
@@ -457,16 +501,3 @@ class AuthObject(object):
             self.handler = self._handlers.get(handler.lower(), urllib2.HTTPBasicAuthHandler)
         else:
             self.handler = handler
-
-class RequestException(Exception):
-    """There was an ambiguous exception that occured while handling your
-    request."""
-
-class AuthenticationError(RequestException):
-    """The authentication credentials provided were invalid."""
-
-class URLRequired(RequestException):
-    """A valid URL is required to make a request."""
-
-class InvalidMethod(RequestException):
-    """An inappropriate method was attempted."""